Validate play trace structure
