+++ dd-smoke-tests/springboot-grpc/src/test/groovy/datadog/smoketest/SpringBootGrpcAsyncConcurrentTest.groovy
@@ -9,8 +9,8 @@ class SpringBootGrpcAsyncConcurrentTest extends SpringBootWithGRPCTest {
       .toSet()
 
   @Override
-  boolean isAcceptable(Map<String, AtomicInteger> traceCounts) {
-    assertTraceCounts(EXPECTED_TRACES, traceCounts)
+  protected Set<String> expectedTraces() {
+    return EXPECTED_TRACES
   }
 
   @Override
