+++ dd-smoke-tests/springboot-grpc/src/test/groovy/datadog/smoketest/SpringBootGrpcSyncTest.groovy
@@ -1,15 +1,13 @@
 package datadog.smoketest
 
-import java.util.concurrent.atomic.AtomicInteger
-
 class SpringBootGrpcSyncTest extends SpringBootWithGRPCTest {
   private static final Set<String> EXPECTED_TRACES =
     ["[grpc.server[grpc.message]]",
             "[servlet.request[spring.handler[grpc.client[grpc.message]]]]"].toSet()
 
   @Override
-  boolean isAcceptable(Map<String, AtomicInteger> traceCounts) {
-    assertTraceCounts(EXPECTED_TRACES, traceCounts)
+  protected Set<String> expectedTraces() {
+    return EXPECTED_TRACES
   }
 
   @Override
